fixed errors with PR
