+++ .changeset/kind-insects-talk.md
@@ -1,0 +1,11 @@
+---
+'@chainlink/1forge-adapter': minor
+'@chainlink/currencylayer-adapter': minor
+'@chainlink/fixer-adapter': minor
+'@chainlink/metalsapi-adapter': minor
+'@chainlink/openexchangerates-adapter': minor
+'@chainlink/polygon-adapter': minor
+'@chainlink/tradermade-adapter': minor
+---
+
+This PR Includes changes for the following adapters: (and their from to to abilities)Polygon: from: many to: manyOpenExchangeRates: from: one to:many1Forge: from: many to: manyTradermade: from: ma
