+++ packages/sources/1forge/src/endpoint/index.ts
@@ -1,3 +1,2 @@
 export * as forex from './forex'
 export * as quotes from './quotes'
-
