+++ packages/sources/1forge/src/endpoint/quotes.ts
@@ -33,7 +33,12 @@ const handleBatchedRequest = (
       Requester.validateResultNumber(pair, [resultPath]),
     ])
   }
-  return Requester.success(jobRunID, Requester.withResult(response, undefined, payload), true, batchablePropertyPath)
+  return Requester.success(
+    jobRunID,
+    Requester.withResult(response, undefined, payload),
+    true,
+    batchablePropertyPath,
+  )
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
