+++ packages/sources/fixer/README.md
@@ -55,23 +55,26 @@ This adapter is for [Fixer.io](https://fixer.io/) and supports the convert endpo
 ```
 
 ## Live Endpoint
+
 #### Returns a batched price comparison from one currency to a list of other currencies.
+
 ### Input Params
 
 | Required? |            Name            |                        Description                        |                                       Options                                        | Defaults to |
 | :-------: | :------------------------: | :-------------------------------------------------------: | :----------------------------------------------------------------------------------: | :---------: |
 |    ✅     | `base`, `from`, or `coin`  |            The symbol of the currency to query            |                                                                                      |             |
-|    ✅     | `quote`, `to`, or `market` |         The symbol of the currencies to convert to               |                                                                                      |      1      |
+|    ✅     | `quote`, `to`, or `market` |        The symbol of the currencies to convert to         |                                                                                      |      1      |
 |    🟡     |        `overrides`         | If base provided is found in overrides, that will be used | [Format](../../core/bootstrap/src/lib/external-adapter/overrides/presetSymbols.json) |             |
 
 ## Input
+
 ```json
 {
-    "id": "1",
-    "data": {
-        "base": "USD",
-        "quote": ["EUR","AUD"]
-    }
+  "id": "1",
+  "data": {
+    "base": "USD",
+    "quote": ["EUR", "AUD"]
+  }
 }
 ```
 
@@ -79,75 +82,71 @@ This adapter is for [Fixer.io](https://fixer.io/) and supports the convert endpo
 
 ```json
 {
-    "jobRunID": "1",
-    "debug": {
-        "staleness": 0,
-        "performance": 0.965477773,
-        "providerCost": 1,
-        "batchablePropertyPath": [
-            "base",
-            "quote"
-        ]
+  "jobRunID": "1",
+  "debug": {
+    "staleness": 0,
+    "performance": 0.965477773,
+    "providerCost": 1,
+    "batchablePropertyPath": ["base", "quote"]
+  },
+  "statusCode": 200,
+  "data": {
+    "success": true,
+    "timestamp": 1519296206,
+    "base": "USD",
+    "date": "2021-07-29",
+    "rates": {
+      "AUD": 1.278342,
+      "EUR": 1.278342,
+      "GBP": 0.908019,
+      "PLN": 3.731504
     },
-    "statusCode": 200,
-    "data": {
-        "success": true,
-        "timestamp": 1519296206,
-        "base": "USD",
-        "date": "2021-07-29",
-        "rates": {
-            "AUD": 1.278342,
-            "EUR": 1.278342,
-            "GBP": 0.908019,
-            "PLN": 3.731504
+    "results": [
+      [
+        {
+          "id": "1",
+          "data": {
+            "base": "USD",
+            "quote": "AUD"
+          },
+          "rateLimitMaxAge": 960
         },
-        "results": [
-            [
-                {
-                    "id": "1",
-                    "data": {
-                        "base": "USD",
-                        "quote": "AUD"
-                    },
-                    "rateLimitMaxAge": 960
-                },
-                1.278342
-            ],
-            [
-                {
-                    "id": "1",
-                    "data": {
-                        "base": "USD",
-                        "quote": "EUR"
-                    },
-                    "rateLimitMaxAge": 960
-                },
-                1.278342
-            ],
-            [
-                {
-                    "id": "1",
-                    "data": {
-                        "base": "USD",
-                        "quote": "GBP"
-                    },
-                    "rateLimitMaxAge": 960
-                },
-                0.908019
-            ],
-            [
-                {
-                    "id": "1",
-                    "data": {
-                        "base": "USD",
-                        "quote": "PLN"
-                    },
-                    "rateLimitMaxAge": 960
-                },
-                3.731504
-            ]
-        ]
-    }
+        1.278342
+      ],
+      [
+        {
+          "id": "1",
+          "data": {
+            "base": "USD",
+            "quote": "EUR"
+          },
+          "rateLimitMaxAge": 960
+        },
+        1.278342
+      ],
+      [
+        {
+          "id": "1",
+          "data": {
+            "base": "USD",
+            "quote": "GBP"
+          },
+          "rateLimitMaxAge": 960
+        },
+        0.908019
+      ],
+      [
+        {
+          "id": "1",
+          "data": {
+            "base": "USD",
+            "quote": "PLN"
+          },
+          "rateLimitMaxAge": 960
+        },
+        3.731504
+      ]
+    ]
+  }
 }
 ```
-
