+++ packages/sources/fixer/src/endpoint/latest.ts
@@ -1,9 +1,15 @@
 import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { ExecuteWithConfig, Config, InputParameters, AdapterRequest, AxiosResponse } from '@chainlink/types'
+import {
+  ExecuteWithConfig,
+  Config,
+  InputParameters,
+  AdapterRequest,
+  AxiosResponse,
+} from '@chainlink/types'
 import { NAME as AdapterName } from '../config'
 
 export const supportedEndpoints = ['latest']
-export const batchablePropertyPath = ['quote']
+export const batchablePropertyPath = [{ name: 'quote' }]
 
 const customError = (data: any) => data.Response === 'Error'
 
@@ -14,25 +20,30 @@ export const inputParameters: InputParameters = {
 }
 
 const handleBatchedRequest = (
-    jobRunID: string,
-    request: AdapterRequest,
-    response: AxiosResponse<any>,
-    resultPath: string,
-    symbols: string[]
-  ) => {
-    const payload: [AdapterRequest, number][] = []
-    for (const symbol of symbols) {
-      const from = response.data.base
-      payload.push([
-        {
-          ...request,
-          data: { ...request.data, base: from.toUpperCase(), quote: symbol.toUpperCase() },
-        },
-        Requester.validateResultNumber(response.data, [resultPath, symbol]),
-      ])
-    }
-    return Requester.success(jobRunID, Requester.withResult(response, undefined, payload), true, batchablePropertyPath)
+  jobRunID: string,
+  request: AdapterRequest,
+  response: AxiosResponse<any>,
+  resultPath: string,
+  symbols: string[],
+) => {
+  const payload: [AdapterRequest, number][] = []
+  for (const symbol of symbols) {
+    const from = response.data.base
+    payload.push([
+      {
+        ...request,
+        data: { ...request.data, base: from.toUpperCase(), quote: symbol.toUpperCase() },
+      },
+      Requester.validateResultNumber(response.data, [resultPath, symbol]),
+    ])
   }
+  return Requester.success(
+    jobRunID,
+    Requester.withResult(response, undefined, payload),
+    true,
+    batchablePropertyPath,
+  )
+}
 
 // NOTE: This endpoint has not been acceptance tested and will need to be once
 // a valid API Key is obtained.
@@ -48,11 +59,10 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
   const params = {
     access_key: config.apiKey,
     base: from,
-    symbols: to.join()
+    symbols: to.join(),
   }
 
   const reqConfig = { ...config.api, params, url }
-
 
   const response = await Requester.request(reqConfig, customError)
   if (Array.isArray(to)) return handleBatchedRequest(jobRunID, request, response, 'rates', to)
