+++ packages/sources/fixer/src/index.ts
@@ -1,5 +1,5 @@
 import { expose } from '@chainlink/ea-bootstrap'
-import { endpointSelector, makeExecute } from './adapter'
+import { makeExecute } from './adapter'
 import { makeConfig, NAME } from './config'
 
 const { server } = expose(NAME, makeExecute())
