+++ packages/sources/tradermade/src/endpoint/forex.ts
@@ -1,6 +1,6 @@
-import { ExecuteWithConfig, Config } from '@chainlink/types'
-import { execute as liveExecute } from './live'
-import { Validator } from '@chainlink/ea-bootstrap'
+import { ExecuteWithConfig, Config, AdapterRequest, AxiosResponse } from '@chainlink/types'
+import { Validator, Requester, util } from '@chainlink/ea-bootstrap'
+import { NAME } from '../config'
 
 /**
  * This endpoint is similar to live but is supposed to only be used to fetch forex data.  This is why quote is a required parameter.
@@ -8,21 +8,67 @@ import { Validator } from '@chainlink/ea-bootstrap'
  */
 
 export const supportedEndpoints = ['forex']
+export const batchablePropertyPath = [{ name: 'from' }, { name: 'to' }]
 
 export const customParams = {
   base: ['base', 'from', 'symbol', 'market'],
   quote: ['quote', 'to', 'market', 'convert'],
 }
 
-export const execute: ExecuteWithConfig<Config> = async (input, context, config) => {
-  const validator = new Validator(input, customParams)
-  if (validator.error) throw validator.error
-  const transformedInputData = {
-    ...input,
-    data: {
-      ...input.data,
-      to: validator.validated.data.quote,
-    },
+const handleBatchedRequest = (
+  jobRunID: string,
+  request: AdapterRequest,
+  response: AxiosResponse<any>,
+  resultPath: string,
+) => {
+  const payload: [AdapterRequest, number][] = []
+  for (const pair of response.data.quotes) {
+    const symbol = pair.base_currency
+    const to = pair.quote_currency
+
+    payload.push([
+      {
+        ...request,
+        data: { ...request.data, from: symbol.toUpperCase(), to: to.toUpperCase() },
+      },
+      Requester.validateResultNumber(pair, [resultPath]),
+    ])
   }
-  return await liveExecute(transformedInputData, context, config)
+  return Requester.success(
+    jobRunID,
+    Requester.withResult(response, undefined, payload),
+    true,
+    batchablePropertyPath,
+  )
+}
+
+export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
+  const validator = new Validator(request, customParams)
+  if (validator.error) throw validator.error
+  Requester.logConfig(config)
+
+  const jobRunID = validator.validated.id
+  const symbol = validator.overrideSymbol(NAME)
+  const to = validator.validated.data.quote || ''
+  const pairArray = []
+
+  for (const fromCurrency of util.formatArray(symbol)) {
+    for (const toCurrency of util.formatArray(to)) {
+      pairArray.push(`${fromCurrency.toUpperCase() + toCurrency.toUpperCase()}`)
+    }
+  }
+
+  const currencies = pairArray.toString()
+  const params = {
+    ...config.api.params,
+    currencies,
+  }
+
+  const options = { ...config.api, params }
+  const response = await Requester.request(options)
+  if (Array.isArray(symbol) || Array.isArray(to))
+    return handleBatchedRequest(jobRunID, request, response, 'mid')
+
+  response.data.result = Requester.validateResultNumber(response.data, ['quotes', 0, 'mid'])
+  return Requester.success(jobRunID, response, config.api.verbose, batchablePropertyPath)
 }
